kernel_10_26_1_18/jb-mr2-t40-r1-10.26.1.18-20131217-1558-foss/mydroid/external/webkit/LayoutTests/platform/android/fast/canvas/webgl/robustness-context-lost-expected.txt

PASS drawing timed out in less than 4 seconds
PASS context is lost
PASS context lost event was received
PASS successfullyParsed is true

TEST COMPLETE
